refactor: move AggregateExpression
